# financeApp-backend-feature-admin-group-management/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance Backend API",
        "description": "RESTful API backend for dual-version finance management application supporting Admin and User versions of Flutter mobile app",
        "contact": {
            "email": "support@financeapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/export/expenses/pdf": {
            "post": {
                "tags": [
                    "Export"
                ],
                "summary": "Export expenses to PDF",
                "description": "Generates a formatted PDF document with expense records and returns a download URL valid for 24 hours",
                "operationId": "b6c6ff0ced8b75e7736fba709d0a4290",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "sync_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "syncing",
                                            "synced",
                                            "failed"
                                        ]
                                    },
                                    "has_invoice": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PDF export generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PDF export generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "export_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                },
                                                "file_name": {
                                                    "type": "string",
                                                    "example": "expenses_export_20251022.pdf"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "download_url": {
                                                    "type": "string",
                                                    "example": "/api/v1/export/1/download"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Export generation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/export/expenses/excel": {
            "post": {
                "tags": [
                    "Export"
                ],
                "summary": "Export expenses to Excel",
                "description": "Generates an XLSX file with expense records in tabular format and returns a download URL valid for 24 hours",
                "operationId": "1d074371c28cef333b598b096b12c1d8",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "sync_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "syncing",
                                            "synced",
                                            "failed"
                                        ]
                                    },
                                    "has_invoice": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Excel export generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Excel export generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "export_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                },
                                                "file_name": {
                                                    "type": "string",
                                                    "example": "expenses_export_20251022.xlsx"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "download_url": {
                                                    "type": "string",
                                                    "example": "/api/v1/export/1/download"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Export generation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/export/{id}/download": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Download export file",
                "description": "Downloads a generated export file. Files are valid for 24 hours after generation.",
                "operationId": "22f3f35fa7d89d94bc2ea33c8879be2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Export ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export file",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Export not ready for download"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Export not found"
                    },
                    "410": {
                        "description": "Export has expired"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}